Move StripedLoader to layout
